requirements.txt
alembic==1.12.1
aniso8601==9.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.22.0
Werkzeug==3.0.1
